Better Cosmic Armor wings
Changes to the boolean isFlying inside ModelArmorInfinity.
The boolean is now true if an EntityPlayer is Flying &&!onGround of isAirBoure.
The Wings will now render after reload if you log out while flying instead of the wing render stopping when you re-log and restarting after you stop flying and touch the ground!