Fallback Mask Texutre and Another bug fix
Fixed another bug that occurs when JEI renders an item with mask opacity set to less than 1.0f
This bug causes another desync with the Infinity Armor Cosmic Model render which is why I have addressed it.

CosmicHaloItemRender and CosmicItemRender now handle proper fallback options for returning Null instead of a TextureAtlasSprite.

MaskOpacity is now set to 1.0f and wont be changed using getMaskOpacity Method. This is the stop the desync with Infinity Armour.
I have kept the method getMaskOpacity to keep inter-compatibility between other mods that use this method.